fix: dkg proto files
