+++ crates/dkg/src/dkgpb/v1/bcast.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package bcast;
+package dkg.dkgpb.v1;
 
 import "google/protobuf/any.proto";
