+++ crates/dkg/src/dkgpb/v1/frost.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package frost;
+package dkg.dkgpb.v1;
 
 message FrostMsgKey {    // dkg.msgKey
   uint32 val_idx = 1;
