--- crates/dkg/src/dkgpb/v1/frost.rs
@@ -1,136 +1,0 @@
-// This file is @generated by prost-build.
-#![allow(dead_code)]
-#![allow(missing_docs)]
-#![allow(clippy::all)]
-#![allow(rustdoc::all)]
-
-/// dkg.msgKey
-#[derive(Clone, Copy, PartialEq, Eq, Hash, ::prost::Message)]
-pub struct FrostMsgKey {
-    #[prost(uint32, tag = "1")]
-    pub val_idx: u32,
-    #[prost(uint32, tag = "2")]
-    pub source_id: u32,
-    #[prost(uint32, tag = "3")]
-    pub target_id: u32,
-}
-impl ::prost::Name for FrostMsgKey {
-    const NAME: &'static str = "FrostMsgKey";
-    const PACKAGE: &'static str = "frost";
-    fn full_name() -> ::prost::alloc::string::String {
-        "frost.FrostMsgKey".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/frost.FrostMsgKey".into()
-    }
-}
-/// Reliable-broadcast
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FrostRound1Casts {
-    /// One per validator
-    #[prost(message, repeated, tag = "1")]
-    pub casts: ::prost::alloc::vec::Vec<FrostRound1Cast>,
-}
-impl ::prost::Name for FrostRound1Casts {
-    const NAME: &'static str = "FrostRound1Casts";
-    const PACKAGE: &'static str = "frost";
-    fn full_name() -> ::prost::alloc::string::String {
-        "frost.FrostRound1Casts".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/frost.FrostRound1Casts".into()
-    }
-}
-#[derive(Clone, PartialEq, Eq, Hash, ::prost::Message)]
-pub struct FrostRound1Cast {
-    #[prost(message, optional, tag = "1")]
-    pub key: ::core::option::Option<FrostMsgKey>,
-    #[prost(bytes = "bytes", tag = "2")]
-    pub wi: ::prost::bytes::Bytes,
-    #[prost(bytes = "bytes", tag = "3")]
-    pub ci: ::prost::bytes::Bytes,
-    #[prost(bytes = "bytes", repeated, tag = "4")]
-    pub commitments: ::prost::alloc::vec::Vec<::prost::bytes::Bytes>,
-}
-impl ::prost::Name for FrostRound1Cast {
-    const NAME: &'static str = "FrostRound1Cast";
-    const PACKAGE: &'static str = "frost";
-    fn full_name() -> ::prost::alloc::string::String {
-        "frost.FrostRound1Cast".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/frost.FrostRound1Cast".into()
-    }
-}
-/// Direct peer-to-peer
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FrostRound1P2p {
-    /// One per validator
-    #[prost(message, repeated, tag = "1")]
-    pub shares: ::prost::alloc::vec::Vec<FrostRound1ShamirShare>,
-}
-impl ::prost::Name for FrostRound1P2p {
-    const NAME: &'static str = "FrostRound1P2P";
-    const PACKAGE: &'static str = "frost";
-    fn full_name() -> ::prost::alloc::string::String {
-        "frost.FrostRound1P2P".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/frost.FrostRound1P2P".into()
-    }
-}
-#[derive(Clone, PartialEq, Eq, Hash, ::prost::Message)]
-pub struct FrostRound1ShamirShare {
-    #[prost(message, optional, tag = "1")]
-    pub key: ::core::option::Option<FrostMsgKey>,
-    #[prost(uint32, tag = "2")]
-    pub id: u32,
-    #[prost(bytes = "bytes", tag = "3")]
-    pub value: ::prost::bytes::Bytes,
-}
-impl ::prost::Name for FrostRound1ShamirShare {
-    const NAME: &'static str = "FrostRound1ShamirShare";
-    const PACKAGE: &'static str = "frost";
-    fn full_name() -> ::prost::alloc::string::String {
-        "frost.FrostRound1ShamirShare".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/frost.FrostRound1ShamirShare".into()
-    }
-}
-/// Reliable-broadcast
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FrostRound2Casts {
-    /// One per validator
-    #[prost(message, repeated, tag = "1")]
-    pub casts: ::prost::alloc::vec::Vec<FrostRound2Cast>,
-}
-impl ::prost::Name for FrostRound2Casts {
-    const NAME: &'static str = "FrostRound2Casts";
-    const PACKAGE: &'static str = "frost";
-    fn full_name() -> ::prost::alloc::string::String {
-        "frost.FrostRound2Casts".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/frost.FrostRound2Casts".into()
-    }
-}
-#[derive(Clone, PartialEq, Eq, Hash, ::prost::Message)]
-pub struct FrostRound2Cast {
-    #[prost(message, optional, tag = "1")]
-    pub key: ::core::option::Option<FrostMsgKey>,
-    #[prost(bytes = "bytes", tag = "2")]
-    pub verification_key: ::prost::bytes::Bytes,
-    #[prost(bytes = "bytes", tag = "3")]
-    pub vk_share: ::prost::bytes::Bytes,
-}
-impl ::prost::Name for FrostRound2Cast {
-    const NAME: &'static str = "FrostRound2Cast";
-    const PACKAGE: &'static str = "frost";
-    fn full_name() -> ::prost::alloc::string::String {
-        "frost.FrostRound2Cast".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/frost.FrostRound2Cast".into()
-    }
-}
