--- crates/dkg/src/dkgpb/v1/nodesigs.rs
@@ -1,23 +1,0 @@
-// This file is @generated by prost-build.
-#![allow(dead_code)]
-#![allow(missing_docs)]
-#![allow(clippy::all)]
-#![allow(rustdoc::all)]
-
-#[derive(Clone, PartialEq, Eq, Hash, ::prost::Message)]
-pub struct MsgNodeSig {
-    #[prost(bytes = "bytes", tag = "1")]
-    pub signature: ::prost::bytes::Bytes,
-    #[prost(uint32, tag = "2")]
-    pub peer_index: u32,
-}
-impl ::prost::Name for MsgNodeSig {
-    const NAME: &'static str = "MsgNodeSig";
-    const PACKAGE: &'static str = "nodesigs";
-    fn full_name() -> ::prost::alloc::string::String {
-        "nodesigs.MsgNodeSig".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/nodesigs.MsgNodeSig".into()
-    }
-}
