+++ crates/dkg/src/dkgpb/v1/sync.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package sync;
+package dkg.dkgpb.v1;
 
 import "google/protobuf/timestamp.proto";
