--- crates/dkg/src/dkgpb/v1/sync.rs
@@ -1,46 +1,0 @@
-// This file is @generated by prost-build.
-#![allow(dead_code)]
-#![allow(missing_docs)]
-#![allow(clippy::all)]
-#![allow(rustdoc::all)]
-
-#[derive(Clone, PartialEq, Eq, Hash, ::prost::Message)]
-pub struct MsgSync {
-    #[prost(message, optional, tag = "1")]
-    pub timestamp: ::core::option::Option<::prost_types::Timestamp>,
-    #[prost(bytes = "bytes", tag = "2")]
-    pub hash_signature: ::prost::bytes::Bytes,
-    #[prost(bool, tag = "3")]
-    pub shutdown: bool,
-    #[prost(string, tag = "4")]
-    pub version: ::prost::alloc::string::String,
-    #[prost(int64, tag = "5")]
-    pub step: i64,
-}
-impl ::prost::Name for MsgSync {
-    const NAME: &'static str = "MsgSync";
-    const PACKAGE: &'static str = "sync";
-    fn full_name() -> ::prost::alloc::string::String {
-        "sync.MsgSync".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/sync.MsgSync".into()
-    }
-}
-#[derive(Clone, PartialEq, Eq, Hash, ::prost::Message)]
-pub struct MsgSyncResponse {
-    #[prost(message, optional, tag = "1")]
-    pub sync_timestamp: ::core::option::Option<::prost_types::Timestamp>,
-    #[prost(string, tag = "2")]
-    pub error: ::prost::alloc::string::String,
-}
-impl ::prost::Name for MsgSyncResponse {
-    const NAME: &'static str = "MsgSyncResponse";
-    const PACKAGE: &'static str = "sync";
-    fn full_name() -> ::prost::alloc::string::String {
-        "sync.MsgSyncResponse".into()
-    }
-    fn type_url() -> ::prost::alloc::string::String {
-        "type.googleapis.com/sync.MsgSyncResponse".into()
-    }
-}
